backend/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
langchain
langchain-google-genai
requests
httpx